move _create_random_clusters to cluster.utils and add 'ch' option
